Push handling of `$DOCKER_CONTENT_TRUST` down into `NewDockerCli`
I'm about to add a new (internal) user of this function and this saves
exporting or duplicating then logic.

Both the `docs/yaml` and `man` use don't do anything which would be affected by
enabling/disabling content trust (nor would you expect them to) so switching
from a hardcoded `false` to probing the envvar makes no practicaly difference.

Note that this changes the API of `NewDockerCli`, but actual external callers
should be unexpected.

Signed-off-by: Ian Campbell <ijc@docker.com>